Update folder naming to try to shorten path.
