+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Security/TokenSecuredServices/TokenSecuredChallenge/App.xaml
