+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Security/TokenSecuredServices/TokenSecuredChallenge/TokenSecuredChallenge.csproj
