+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Security/TokenSecuredServices/TokenSecuredKnownUser/TokenSecuredKnownUser.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.40629.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TokenSecuredServicesKnownUser", "TokenSecuredServicesKnownUser.csproj", "{39DE280E-2709-4C05-8F0C-01F8774D73C7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TokenSecuredKnownUser", "TokenSecuredKnownUser.csproj", "{39DE280E-2709-4C05-8F0C-01F8774D73C7}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8ED45222-07B5-4E0C-8D2B-E1686D463A46}"
 	ProjectSection(SolutionItems) = preProject
