+++ src/Store/ArcGISRuntimeSamplesStore/Samples/Security/TokenSecuredServices/TokenSecuredChallenge/App.xaml.cs
