+++ src/Store/ArcGISRuntimeSamplesStore/Samples/Security/TokenSecuredServices/TokenSecuredChallenge/MainPage.xaml
