+++ src/Store/ArcGISRuntimeSamplesStore/Samples/Security/TokenSecuredServices/TokenSecuredChallenge/Package.appxmanifest
