+++ src/Store/ArcGISRuntimeSamplesStore/Samples/Security/TokenSecuredServices/TokenSecuredChallenge/TokenSecuredChallenge_TemporaryKey.pfx
