+++ src/Store/ArcGISRuntimeSamplesStore/Samples/Security/TokenSecuredServices/TokenSecuredKnownUser/Assets/SplashScreen.png
