+++ src/Store/ArcGISRuntimeSamplesStore/Samples/Security/TokenSecuredServices/TokenSecuredKnownUser/Assets/StoreLogo.png
