+++ src/Store/ArcGISRuntimeSamplesStore/Samples/Security/TokenSecuredServices/TokenSecuredKnownUser/TokenSecuredKnownUser_TemporaryKey.pfx
